feat: get token and save methods

+++ pages/api/items/[id].js
@@ -1,6 +1,9 @@
 import { getItem, updateItem, editItem, deleteItem } from "@/helpers/db";
+import { getToken } from "next-auth/jwt";
 
 export default async function handler(request, response) {
+  const token = await getToken({ req: request });
+
   switch (request.method) {
     case "GET": {
       const item = await getItem(request.query.id);
@@ -16,44 +19,50 @@ export default async function handler(request, response) {
     }
 
     case "PUT": {
-      const item = JSON.parse(request.body);
-      const updatedItem = await updateItem(request.query.id, item);
-      if (!updatedItem) {
-        response.status(404).json({
-          message: `Item ${request.query.id} was not found.`,
-        });
-        return;
-      }
+      if (token) {
+        const item = JSON.parse(request.body);
+        const updatedItem = await updateItem(request.query.id, item);
+        if (!updatedItem) {
+          response.status(404).json({
+            message: `Item ${request.query.id} was not found.`,
+          });
+          return;
+        }
 
-      response.status(200).json(updatedItem);
-      break;
+        response.status(200).json(updatedItem);
+        break;
+      }
     }
 
     case "PATCH": {
-      const item = JSON.parse(request.body);
-      const editedItem = await editItem(request.query.id, item);
-      if (!editedItem) {
-        response.status(404).json({
-          message: `Item ${request.query.id} was not found.`,
-        });
-        return;
-      }
+      if (token) {
+        const item = JSON.parse(request.body);
+        const editedItem = await editItem(request.query.id, item);
+        if (!editedItem) {
+          response.status(404).json({
+            message: `Item ${request.query.id} was not found.`,
+          });
+          return;
+        }
 
-      response.status(200).json(editedItem);
-      break;
+        response.status(200).json(editedItem);
+        break;
+      }
     }
 
     case "DELETE": {
-      const deletedItem = await deleteItem(request.query.id);
-      if (!deletedItem) {
-        response.status(404).json({
-          message: `Item ${request.query.id} was not found.`,
-        });
-        return;
-      }
+      if (token) {
+        const deletedItem = await deleteItem(request.query.id);
+        if (!deletedItem) {
+          response.status(404).json({
+            message: `Item ${request.query.id} was not found.`,
+          });
+          return;
+        }
 
-      response.status(200).json(deletedItem);
-      break;
+        response.status(200).json(deletedItem);
+        break;
+      }
     }
 
     default: {
